requirements.txt
Django>5.1
Pillow
djangorestframework
djangorestframework-api-key>=3.0
django-taggit